Reorganize parser into a Parsing folder
