+++ src/Spectre.Console.Ansi/Parsing/AnsiParserState.cs
@@ -1,0 +1,19 @@
+namespace Spectre.Console.Ansi;
+
+internal enum AnsiParserState
+{
+    Ground = 0,
+    Escape,
+    EscapeIntermediate,
+    CsiEntry,
+    CsiIntermediate,
+    CsiParam,
+    CsiIgnore,
+    DcsEntry,
+    DcsParam,
+    DcsIntermediate,
+    DcsPassthrough,
+    DcsIgnore,
+    OscString,
+    SosPmApcString,
+}
