+++ src/Spectre.Console.Ansi/Parsing/AnsiToken.cs
