+++ src/Spectre.Console.Ansi/Parsing/AnsiTransition.cs
@@ -1,0 +1,5 @@
+namespace Spectre.Console.Ansi;
+
+internal readonly record struct AnsiTransition(
+    AnsiParserState State,
+    AnsiTransitionAction Action);
