+++ src/Spectre.Console.Ansi/Parsing/AnsiTransitionAction.cs
@@ -1,0 +1,15 @@
+namespace Spectre.Console.Ansi;
+
+internal enum AnsiTransitionAction
+{
+    None = 0,
+    Ignore,
+    Print,
+    Execute,
+    Collect,
+    Param,
+    EscDispatch,
+    CsiDispatch,
+    OscPut,
+    DscPut,
+}
