+++ src/Spectre.Console.Ansi/Parsing/Osc/OscCommand.cs
