+++ src/Spectre.Console.Ansi/Parsing/Osc/OscHyperLinkParser.cs
