+++ src/Spectre.Console.Ansi/Parsing/Osc/OscParser.cs
